refactor(migration): group drop-in shims under turbohtml.migration
Move the compatibility layers out of the package root into a migration
subpackage named after the library each one replaces: bleach_compat.py ->
migration/bleach.py, html_parser.py -> migration/stdlib.py (html.parser
drop-in), markup.py -> migration/markupsafe.py. Root now carries only
turbohtml's own modules. Bumps relative imports, meson sources, README
examples, and the affected test imports.
