+++ src/turbohtml/migration/stdlib.py
@@ -23,10 +23,10 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, cast
 
-from ._html import Tokenizer, TokenType
+from .._html import Tokenizer, TokenType
 
 if TYPE_CHECKING:
-    from ._html import Token
+    from .._html import Token
 
 __all__ = ["HTMLParser"]
